Refactor Sow::Source
- remove Data namespace
