+++ lib/sow/parser/json.rb
@@ -1,0 +1,9 @@
+module Sow
+  module Parser
+    class Json < Base
+      def parse
+        JSON.load(data_io)
+      end
+    end
+  end
+end
