+++ lib/sow/parser/yml.rb
@@ -1,0 +1,9 @@
+module Sow
+  module Parser
+    class Yml < Base
+      def parse
+        YAML.load(data_io)
+      end
+    end
+  end
+end
