+++ spec/lib/sow/directive_spec.rb
@@ -44,31 +44,13 @@ describe Sow::Directive do
     end
   end
 
-  describe "#data" do
-    let(:datasource) { double('datasource') }
-    let(:data_hash)  { double('data_hash') }
-
-    subject { described_class.new(:class => Post, :source => 'source') }
-
-    before do
-      datasource.stub_chain(:to_hash, :with_indifferent_access).and_return(data_hash)
-
-      Sow::Data::Source.stub(:new).with('posts', { :source => 'source' }).and_return(datasource)
-    end
-
-    it "returns data from the datasource" do
-      subject.data.should == data_hash
-    end
-
-  end
-
   describe "#datasource" do
     let(:datasource) { double('datasource') }
 
     subject { described_class.new(:class => Post, :source => 'source') }
 
     before do
-      Sow::Data::Source.stub(:new).with('posts', { :source => 'source' }).and_return(datasource)
+      Sow::Source.stub(:new).with('posts', { :source => 'source' }).and_return(datasource)
     end
 
     it "returns a sow data source" do
